Bug fixes, code smell fixes, and duplication code fix

+++ sm-core/src/main/java/com/salesmanager/core/business/modules/integration/shipping/impl/CustomShippingQuoteRules.java
@@ -103,8 +103,8 @@ public class CustomShippingQuoteRules implements ShippingQuoteModule {
 			sizeList.add(pack.getShippingWeight());
 			sizeList.add(pack.getShippingLength());
 			Double maxSize = (Double)Collections.max(sizeList);
-			if(size==null || maxSize.doubleValue() > size.doubleValue()) {
-				size = maxSize.doubleValue();
+			if(size==null || maxSize > size) {
+				size = maxSize;
 			}
 		}
