+++ sm-core/src/main/java/com/salesmanager/core/business/modules/integration/shipping/impl/ShippingDecisionPreProcessorImpl.java
@@ -100,8 +100,8 @@ public class ShippingDecisionPreProcessorImpl implements ShippingQuotePrePostPro
 			sizeList.add(pack.getShippingLength());
 			sizeList.add(pack.getShippingWidth());
 			Double maxSize = (Double)Collections.max(sizeList);
-			if(size==null || maxSize.doubleValue() > size.doubleValue()) {
-				size = maxSize.doubleValue();
+			if(size==null || maxSize > size) {
+				size = maxSize;
 			}
 		}
